+++ src/components/productCardList/ProductCardList.tsx
@@ -16,7 +16,7 @@ interface ProductCardListProps {
     React.SetStateAction<Record<"productId" | "cartId", number>[]>
   >;
   setErrorTrue: (type: ERROR_TYPE) => void;
-  fetchProducts: () => void;
+  fetchCartProducts: () => void;
 }
 
 function ProductCardList({
@@ -27,7 +27,7 @@ function ProductCardList({
   cartItemIds,
   setCartItemIds,
   setErrorTrue,
-  fetchProducts,
+  fetchCartProducts,
 }: ProductCardListProps) {
   const { isLoading } = useFetchProducts({
     category,
@@ -53,7 +53,7 @@ function ProductCardList({
           imageUrl={data.imageUrl}
           setCartItemIds={setCartItemIds}
           setErrorTrue={setErrorTrue}
-          fetchProducts={fetchProducts}
+          fetchCartProducts={fetchCartProducts}
         />
       ))}
     </div>
